chore(refactor): change package name
