+++ pyproject.toml
@@ -1,5 +1,5 @@
 [tool.poetry]
-name = "dbkit"
+name = "sanctumlabs-dbkit"
 # This version field is not being used. The version field is dynamically updated as below
 version = "0.0.0"
 description = "A wrapper for database libraries which bakes in common database functionality."
