+++ sanctumlabs_dbkit/exceptions.py
