+++ sanctumlabs_dbkit/protocols.py
@@ -1,6 +1,7 @@
 """
 Protocols defining types
 """
+
 from typing import Protocol
