+++ sanctumlabs_dbkit/sql/types.py
@@ -1,7 +1,8 @@
 """
 Database Kit Types
 """
+
 from typing import Callable
-from dbkit.sql.session import Session
+from sanctumlabs_dbkit.sql.session import Session
 
 CommitCallback = Callable[[Session], None]
