+++ sanctumlabs_dbkit/utils.py
@@ -1,9 +1,10 @@
 """
 Database utilities
 """
+
 from typing import Optional
-from dbkit.protocols import User
-from dbkit.sql.models import AbstractBaseModel
+from sanctumlabs_dbkit.protocols import User
+from sanctumlabs_dbkit.sql.models import AbstractBaseModel
 
 
 def set_updated_by(user: Optional[User], entity: AbstractBaseModel) -> None:
